requirements.txt
psycopg==3.1.17
psycopg2-binary==2.9.9
pyTelegramBotAPI==4.14.1
telebot==0.0.5